requirements.txt
Django==5.0.6
djangorestframework==3.15.1
pillow==10.3.0
psycopg2-binary
python-dotenv==1.0.1
django-filter
djangorestframework-simplejwt==5.3.1
coverage
setuptools
django-cors-headers
requests
drf_yasg
stripe
celery
django-celery-beat
eventlet
redis
pytz